backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# 数据库（支持 PostgreSQL 与 MySQL）
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
# MySQL 异步驱动（本地用 MySQL 时必装）
asyncmy==0.2.9
pymysql==1.1.0

# Redis（hiredis 2.2.0 在 PyPI 已不可用，用 2.2.1+）
redis==5.0.1
hiredis>=2.2.1

# 认证（密码哈希直接使用 bcrypt，不依赖 passlib）
python-jose[cryptography]==3.3.0
bcrypt>=4.0.1
python-multipart==0.0.6

# 文件处理
PyPDF2==3.0.1
pdfplumber==0.10.3
pdf2image==1.16.3
python-pptx==0.6.23
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.3
Pillow==10.1.0
opencv-python==4.8.1.78

# OCR（可选：部分环境无 paddlepaddle 对应 wheel，需 OCR 时再单独安装）
# paddlepaddle、paddleocr 已移出默认依赖，安装命令: pip install paddlepaddle paddleocr

# AI模型
openai==1.3.7
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1

# 向量数据库（支持 Qdrant / Zilliz Cloud）
qdrant-client==1.7.0
pymilvus==2.4.0
# 注意：pymilvus 需要 pkg_resources，Python 3.12 中需要 setuptools<69.0.0
# marshmallow 4.x 移除了 __version_info__，pymilvus 需要 3.x
marshmallow>=3.20.0,<4.0.0

# 对象存储
minio==7.2.0
boto3==1.29.7

# 任务队列
celery==5.3.4
flower==2.0.1

# HTTP客户端
httpx==0.25.1
aiohttp==3.9.1

# 工具库
# Python 3.12 兼容性：setuptools>=68.0.0 移除了 pkg_resources，pymilvus 需要它
# 方案1：降级 setuptools 到包含 pkg_resources 的版本（推荐）
setuptools>=65.5.0,<68.0.0  # 提供 pkg_resources 模块（pymilvus 等依赖需要）
# 方案2：如果必须使用 setuptools>=68.0.0，可以尝试安装 pkg_resources-py310（但可能不兼容）
# pkg-resources-py310  # 仅在 setuptools>=68.0.0 时使用
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# 日志
loguru==0.7.2

# 监控
prometheus-client==0.19.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# 代码质量
black==23.11.0
flake8==6.1.0
mypy==1.7.1
